requirements.txt
fastapi
uvicorn
pillow
numpy
python-multipart
tensorflow==2.13.*

